Moved the selection logic to a plugin
Implements #2041
